+++ addons/es/verifactu/scenarios_test.go
@@ -42,15 +42,25 @@ func TestInvoiceDocumentScenarios(t *testing.T) {
 		assert.Equal(t, i.Tax.Ext.Get(verifactu.ExtKeyCorrectionType).String(), "S")
 	})
 
-	t.Run("simplified corrective", func(t *testing.T) {
+	t.Run("corrective special", func(t *testing.T) {
+		i := testInvoiceStandard(t)
+		require.NoError(t, i.Calculate())
+		require.NoError(t, i.Correct(bill.Corrective, bill.WithExtension(verifactu.ExtKeyDocType, "R2")))
+		assert.Equal(t, "R2", i.Tax.Ext.Get(verifactu.ExtKeyDocType).String())
+	})
+
+	t.Run("simplified correcitve invoice", func(t *testing.T) {
 		i := testInvoiceStandard(t)
 		i.SetTags(tax.TagSimplified)
 		require.NoError(t, i.Calculate())
+		require.NoError(t, i.Correct(bill.Corrective))
+		assert.Equal(t, i.Tax.Ext[verifactu.ExtKeyDocType].String(), "R5")
+	})
 
-		require.NoError(t, i.Correct(bill.Corrective, bill.WithExtension(verifactu.ExtKeyDocType, "F3")))
-
-		assert.Len(t, i.Notes, 1)
-		assert.Equal(t, "F3", i.Tax.Ext.Get(verifactu.ExtKeyDocType).String())
-		assert.Equal(t, "", i.Tax.Ext.Get(verifactu.ExtKeyCorrectionType).String())
+	t.Run("replacement", func(t *testing.T) {
+		i := testInvoiceStandard(t)
+		i.SetTags(tax.TagReplacement)
+		require.NoError(t, i.Calculate())
+		assert.Equal(t, i.Tax.Ext[verifactu.ExtKeyDocType].String(), "F3")
 	})
 }