+++ tax/constants.go
@@ -28,6 +28,7 @@ const (
 	TagReverseCharge cbc.Key = "reverse-charge"
 	TagCustomerRates cbc.Key = "customer-rates"
 	TagSelfBilled    cbc.Key = "self-billed"
+	TagReplacement   cbc.Key = "replacement"
 	TagPartial       cbc.Key = "partial"
 	TagB2G           cbc.Key = "b2g"
 	TagExport        cbc.Key = "export"